pyproject.toml
[tool.poetry]
name = "rcg_app_2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-SQLAlchemy = "^2.5.1"
spotipy = "^2.20.0"
wikipedia = "^1.4.0"
pylast = "^5.0.0"
flask-marshmallow = "^0.14.0"
connexion = "^2.14.1"
marshmallow-sqlalchemy = "^0.28.1"
psycopg2 = "^2.9.4"
gunicorn = "^20.1.0"
Jinja2 = "^3.1.2"
click = "^8.1.3"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
